// src/utils/booking-type.ts
    export interface IPriceComponent {
        id?: any;
        listingFeeSettingId?: any;
        type: string;
        name: string;
        title: string;
        alias?: any;
        quantity?: any;
        value: number;
        total: number;
        isIncludedInTotalPrice: number;
        isOverriddenByUser: number;
        isQuantitySelectable: number;
        isMandatory?: any;
        isDeleted: number;
        displayInRent: number;
    }

    export interface IPrice {
        totalPrice: number;
        originalTotalPrice: number;
        selectedComponentsAmount: number;
        components: IPriceComponent[];
        extraComponents: any[];
        selectedComponents: any[];
    }

    export interface HostawayResponse {
        listingMapId: number;
        listingName: string;
        channelId: number;
        source?: any;
        channelName: string;
        reservationId: string;
        hostawayReservationId: string;
        channelReservationId: string;
        externalPropertyId?: any;
        externalUnitId?: any;
        assigneeUserId?: any;
        customerIcalId?: any;
        customerIcalName?: any;
        guestAuthHash: string;
        guestPortalUrl: string;
        guestPortalRevampUrl?: any;
        isProcessed: number;
        isInitial: number;
        isManuallyChecked: number;
        isInstantBooked: number;
        reservationDate: string;
        pendingExpireDate?: any;
        guestName: string;
        guestFirstName?: any;
        guestLastName?: any;
        guestExternalAccountId?: any;
        guestZipCode?: any;
        guestAddress?: any;
        guestCity?: any;
        guestCountry?: any;
        guestEmail: string;
        guestPicture?: any;
        guestRecommendations: number;
        guestTrips: number;
        guestWork?: any;
        isGuestIdentityVerified: number;
        isGuestVerifiedByEmail: number;
        isGuestVerifiedByWorkEmail: number;
        isGuestVerifiedByFacebook: number;
        originalChannel?: any;
        isGuestVerifiedByGovernmentId: number;
        isGuestVerifiedByPhone: number;
        isGuestVerifiedByReviews: number;
        numberOfGuests: number;
        adults?: any;
        children?: any;
        infants?: any;
        pets?: any;
        arrivalDate: string;
        departureDate: string;
        isDatesUnspecified: number;
        previousArrivalDate?: any;
        previousDepartureDate?: any;
        checkInTime: number;
        checkOutTime: number;
        nights: number;
        phone: string;
        totalPrice: number;
        taxAmount: number;
        channelCommissionAmount?: any;
        hostawayCommissionAmount?: any;
        cleaningFee: number;
        securityDepositFee: number;
        isPaid: number;
        ccName?: any;
        ccNumber?: any;
        ccNumberEndingDigits?: any;
        ccExpirationYear?: any;
        ccExpirationMonth?: any;
        cvc?: any;
        stripeGuestId?: any;
        stripeMessage?: any;
        braintreeGuestId?: any;
        braintreeMessage?: any;
        currency: string;
        status: string;
        cancellationDate?: any;
        cancelledBy?: any;
        hostNote?: any;
        guestNote: string;
        doorCode?: any;
        doorCodeVendor?: any;
        doorCodeInstruction?: any;
        comment?: any;
        confirmationCode?: any;
        airbnbExpectedPayoutAmount?: any;
        airbnbListingBasePrice?: any;
        airbnbListingCancellationHostFee?: any;
        airbnbListingCancellationPayout?: any;
        airbnbListingCleaningFee?: any;
        airbnbListingHostFee?: any;
        airbnbListingSecurityPrice?: any;
        airbnbOccupancyTaxAmountPaidToHost?: any;
        airbnbTotalPaidAmount?: any;
        airbnbTransientOccupancyTaxPaidAmount?: any;
        airbnbCancellationPolicy?: any;
        isStarred: number;
        isArchived: number;
        isPinned: number;
        reservationCouponId?: any;
        customFieldValues: any[];
        reservationFees: any[];
        reservationUnit: any[];
        insertedOn: string;
        updatedOn: string;
        latestActivityOn: string;
        customerUserId?: any;
        guestLocale?: any;
        localeForMessaging?: any;
        localeForMessagingSource?: any;
        listingCustomFields?: any;
        rentalAgreementFileUrl?: any;
        reservationAgreement?: any;
        financeField: any[];
        id: number;
    }

    export interface IBooking {
        bookingDate: Date;
        arrivalDate: Date;
        departureDate: Date;
        totalNights: number;
        price: IPrice;
        syncedWithHostaway: boolean;
        hostawayReservationId: string;
        createdAt: Date;
        updatedAt: Date;
        hostawayResponse: HostawayResponse;
        numberOfGuests: number;
        couponName?: string;
    }